requirements.txt
fastapi
uvicorn
PyMySQL
python-jose[cryptography]
bcrypt
python-dotenv